# frontend/contracts/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sadhana Card Tracker API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/corrections/devotees/{user_id}/cards/{card_date}": {
      "get": {
        "tags": [
          "admin-corrections"
        ],
        "summary": "Admin Get Historical Card",
        "operationId": "admin_get_historical_card_api_v1_admin_corrections_devotees__user_id__cards__card_date__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "card_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Card Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyCardPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin-corrections"
        ],
        "summary": "Admin Correct Historical Card",
        "operationId": "admin_correct_historical_card_api_v1_admin_corrections_devotees__user_id__cards__card_date__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "card_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Card Date"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalCardCorrectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalCardCorrectionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cards/today": {
      "get": {
        "tags": [
          "daily-cards"
        ],
        "summary": "Today Card",
        "operationId": "today_card_api_v1_cards_today_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyCardPublic"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "daily-cards"
        ],
        "summary": "Update Today Card",
        "operationId": "update_today_card_api_v1_cards_today_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyCardUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyCardPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/cards/{card_date}": {
      "get": {
        "tags": [
          "daily-cards"
        ],
        "summary": "Dated Card",
        "operationId": "dated_card_api_v1_cards__card_date__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "card_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Card Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyCardPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lifecycle/status": {
      "get": {
        "tags": [
          "devotee-lifecycle"
        ],
        "summary": "Lifecycle Status",
        "operationId": "lifecycle_status_api_v1_lifecycle_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleStatusPublic"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/lifecycle/bhima-status": {
      "post": {
        "tags": [
          "devotee-lifecycle"
        ],
        "summary": "Change Bhima Status",
        "operationId": "change_bhima_status_api_v1_lifecycle_bhima_status_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BhimaStatusChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleStatusPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/weekly/latest": {
      "get": {
        "tags": [
          "weekly-evaluations"
        ],
        "summary": "Latest Weekly Evaluation",
        "operationId": "latest_weekly_evaluation_api_v1_weekly_latest_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyEvaluationPublic"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/weekly/{week_start_date}": {
      "get": {
        "tags": [
          "weekly-evaluations"
        ],
        "summary": "Weekly Evaluation",
        "operationId": "weekly_evaluation_api_v1_weekly__week_start_date__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "week_start_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Week Start Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyEvaluationPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monthly/latest": {
      "get": {
        "tags": [
          "monthly-reports"
        ],
        "summary": "Latest Monthly Report",
        "operationId": "latest_monthly_report_api_v1_monthly_latest_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FourWeekReportPublic"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/monthly/{period_start_date}": {
      "get": {
        "tags": [
          "monthly-reports"
        ],
        "summary": "Monthly Report",
        "operationId": "monthly_report_api_v1_monthly__period_start_date__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period_start_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Period Start Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FourWeekReportPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BhimaStatusChangeRequest": {
        "properties": {
          "employment_status": {
            "$ref": "#/components/schemas/EmploymentStatus"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason",
            "default": "Devotee updated Bhima employment status"
          }
        },
        "type": "object",
        "required": [
          "employment_status"
        ],
        "title": "BhimaStatusChangeRequest"
      },
      "CardStatus": {
        "type": "string",
        "enum": [
          "IN_PROGRESS",
          "FINALIZED"
        ],
        "title": "CardStatus"
      },
...
      "DailyActivityUpdate": {
        "properties": {
          "activity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Activity Id"
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/DailyActivityValueUpdate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "activity_id",
          "values"
        ],
        "title": "DailyActivityUpdate"
      },
...
      "DailyCardPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "card_date": {
            "type": "string",
            "format": "date",
            "title": "Card Date"
          },
          "status": {
            "$ref": "#/components/schemas/CardStatus"
          },
          "editable": {
            "type": "boolean",
            "title": "Editable"
          },
          "category_code": {
            "type": "string",
            "title": "Category Code"
          },
          "category_name": {
            "type": "string",
            "title": "Category Name"
          },
          "first_update_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Update At"
          },
          "last_update_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Update At"
          },
          "finalized_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalized At"
          },
          "timezone_snapshot": {
            "type": "string",
            "title": "Timezone Snapshot"
          },
          "deadline_time_snapshot": {
            "type": "string",
            "format": "time",
            "title": "Deadline Time Snapshot"
          },
          "deadline_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Deadline At Utc"
          },
          "revision_number": {
            "type": "integer",
            "title": "Revision Number"
          },
          "activities": {
            "items": {
              "$ref": "#/components/schemas/DailyActivityPublic"
            },
            "type": "array",
            "title": "Activities"
          }
        },
        "type": "object",
        "required": [
          "id",
          "card_date",
          "status",
          "editable",
          "category_code",
          "category_name",
          "first_update_at",
          "last_update_at",
          "finalized_at",
          "timezone_snapshot",
          "deadline_time_snapshot",
          "deadline_at_utc",
          "revision_number",
          "activities"
        ],
        "title": "DailyCardPublic"
      },
      "DailyCardUpdateRequest": {
        "properties": {
          "activities": {
            "items": {
              "$ref": "#/components/schemas/DailyActivityUpdate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Activities"
          }
        },
        "type": "object",
        "required": [
          "activities"
        ],
        "title": "DailyCardUpdateRequest"
      },
...
      "EmploymentStatus": {
        "type": "string",
        "enum": [
          "WORKING",
          "NOT_WORKING"
        ],
        "title": "EmploymentStatus"
      },
...
      "FourWeekReportPublic": {
        "properties": {
          "period_start_date": {
            "type": "string",
            "format": "date",
            "title": "Period Start Date"
          },
          "period_end_date": {
            "type": "string",
            "format": "date",
            "title": "Period End Date"
          },
          "weeks_count": {
            "type": "integer",
            "title": "Weeks Count"
          },
          "definition": {
            "type": "string",
            "title": "Definition",
            "default": "FOUR_ADJACENT_COMPLETE_WEEKS"
          },
          "weeks": {
            "items": {
              "$ref": "#/components/schemas/FourWeekWeekPublic"
            },
            "type": "array",
            "title": "Weeks"
          },
          "sadhana": {
            "$ref": "#/components/schemas/FourWeekScoreSummaryPublic"
          },
          "academic": {
            "$ref": "#/components/schemas/FourWeekScoreSummaryPublic"
          },
          "activities": {
            "items": {
              "$ref": "#/components/schemas/FourWeekActivityPublic"
            },
            "type": "array",
            "title": "Activities"
          },
          "previous_period": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviousFourWeekSummaryPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "observations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Observations"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "period_start_date",
          "period_end_date",
          "weeks_count",
          "sadhana",
          "academic"
        ],
        "title": "FourWeekReportPublic",
        "description": "Project 'monthly' report: exactly four adjacent complete organization weeks."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoricalCardCorrectionRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason"
          },
          "activities": {
            "items": {
              "$ref": "#/components/schemas/DailyActivityUpdate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Activities"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "activities"
        ],
        "title": "HistoricalCardCorrectionRequest",
        "description": "Admin-only correction of raw values on an already-finalized card.\n\nThe request intentionally reuses the typed DailyActivityUpdate contract so missing\nversus intentional zero/False semantics remain identical to normal devotee updates."
      },
      "HistoricalCardCorrectionResult": {
        "properties": {
          "devotee_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Devotee User Id"
          },
          "card_date": {
            "type": "string",
            "format": "date",
            "title": "Card Date"
          },
          "card": {
            "$ref": "#/components/schemas/DailyCardPublic"
          },
          "weekly_evaluation_recalculated": {
            "type": "boolean",
            "title": "Weekly Evaluation Recalculated"
          },
          "weekly_evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeeklyEvaluationPublic"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "devotee_user_id",
          "card_date",
          "card",
          "weekly_evaluation_recalculated"
        ],
        "title": "HistoricalCardCorrectionResult"
      },
...
      "LifecycleStatusPublic": {
        "properties": {
          "current_category_code": {
            "type": "string",
            "title": "Current Category Code"
          },
          "current_category_name": {
            "type": "string",
            "title": "Current Category Name"
          },
          "current_academic_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Academic Year"
          },
          "next_promotion_week": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Promotion Week"
          },
          "sahadeva_review_due": {
            "type": "boolean",
            "title": "Sahadeva Review Due"
          },
          "bhima_choice_required": {
            "type": "boolean",
            "title": "Bhima Choice Required"
          },
          "pending_category_transition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingCategoryTransitionPublic"
              },
              {
                "type": "null"
              }
            ]
          },
          "pending_deactivation_week": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Deactivation Week"
          },
          "pending_deactivation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Deactivation Reason"
          }
        },
        "type": "object",
        "required": [
          "current_category_code",
          "current_category_name",
          "current_academic_year",
          "next_promotion_week",
          "sahadeva_review_due",
          "bhima_choice_required",
          "pending_category_transition",
          "pending_deactivation_week",
          "pending_deactivation_reason"
        ],
        "title": "LifecycleStatusPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WeeklyEvaluationPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "week_start_date": {
            "type": "string",
            "format": "date",
            "title": "Week Start Date"
          },
          "week_end_date": {
            "type": "string",
            "format": "date",
            "title": "Week End Date"
          },
          "category_code": {
            "type": "string",
            "title": "Category Code"
          },
          "category_name": {
            "type": "string",
            "title": "Category Name"
          },
          "timezone_snapshot": {
            "type": "string",
            "title": "Timezone Snapshot"
          },
          "sadhana_score": {
            "type": "integer",
            "title": "Sadhana Score"
          },
          "sadhana_max_score": {
            "type": "integer",
            "title": "Sadhana Max Score"
          },
          "sadhana_percentage": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sadhana Percentage"
          },
          "academic_score": {
            "type": "integer",
            "title": "Academic Score"
          },
          "academic_max_score": {
            "type": "integer",
            "title": "Academic Max Score"
          },
          "academic_percentage": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Academic Percentage"
          },
          "revision_number": {
            "type": "integer",
            "title": "Revision Number"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "activities": {
            "items": {
              "$ref": "#/components/schemas/WeeklyActivityResultPublic"
            },
            "type": "array",
            "title": "Activities"
          },
          "previous_week": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviousWeekEvaluationPublic"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "week_start_date",
          "week_end_date",
          "category_code",
          "category_name",
          "timezone_snapshot",
          "sadhana_score",
          "sadhana_max_score",
          "academic_score",
          "academic_max_score",
          "revision_number",
          "generated_at",
          "updated_at"
        ],
        "title": "WeeklyEvaluationPublic"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}